refactor(controllers): extract HibernationManager from HelpSessionController
- Move idle-hibernate arm/fire with busy-recheck, system-suspend tracking,
  and the main-captured pending-hibernation seed into a sibling module
- Controller delegates through a narrow host interface (getSnapshot/patch/
  resetPhase/isLaunchCurrent); _applyStopSuppression now calls a single
  clearAndSuppress(projectId) instead of reaching into moved-away fields
- Update the two HelpSessionController.test.ts tests that reach into
  hibernate internals to follow the new _hibernationManager location
- No behavior change
